requirements.txt
# Runtime dependencies (CPU PyTorch is installed separately in Dockerfile).
# Pin only where needed for binary compatibility.

fastapi
uvicorn
pydantic
PyYAML

numpy==1.26.4
opencv-python-headless==4.8.1.78
ultralytics

aiortc
av
firebase_admin
